create PostHeader components
